pipeline: change pipeline_layout.nb_{textures,uniforms,buffers,attributes} types to size_t

+++ libnopegl/src/pgcraft.c
@@ -1572,13 +1572,13 @@ struct pipeline_layout ngli_pgcraft_get_pipeline_layout(const struct pgcraft *s)
 {
     const struct pipeline_layout layout = {
         .uniforms_desc   = ngli_darray_data(&s->filtered_pipeline_info.desc.uniforms),
-        .nb_uniforms     = (int)ngli_darray_count(&s->filtered_pipeline_info.desc.uniforms),
+        .nb_uniforms     = ngli_darray_count(&s->filtered_pipeline_info.desc.uniforms),
         .textures_desc   = ngli_darray_data(&s->filtered_pipeline_info.desc.textures),
-        .nb_textures     = (int)ngli_darray_count(&s->filtered_pipeline_info.desc.textures),
+        .nb_textures     = ngli_darray_count(&s->filtered_pipeline_info.desc.textures),
         .attributes_desc = ngli_darray_data(&s->filtered_pipeline_info.desc.attributes),
-        .nb_attributes   = (int)ngli_darray_count(&s->filtered_pipeline_info.desc.attributes),
+        .nb_attributes   = ngli_darray_count(&s->filtered_pipeline_info.desc.attributes),
         .buffers_desc    = ngli_darray_data(&s->filtered_pipeline_info.desc.buffers),
-        .nb_buffers      = (int)ngli_darray_count(&s->filtered_pipeline_info.desc.buffers),
+        .nb_buffers      = ngli_darray_count(&s->filtered_pipeline_info.desc.buffers),
     };
     return layout;
 }
